+++ src/components/common/Modal/ModalBase/ModalManger.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import ModalBase from "./ModalBase";
-import ModalWithPagination from "./ModalWithPagination";
+import ModalWithPagination from "../ModalWithPagination/ModalWithPagination";
 import { useModalContext } from "./ModalContext";
 
 export default function ModalManager() {
