+++ src/components/common/Modal/ModalWithPagination/ModalHeader.tsx
@@ -1,0 +1,23 @@
+import React from "react";
+
+export default function ModalHeader({
+  title,
+  closeModal,
+}: ModalHeaderProps) {
+  return (
+    <div className="flex justify-between items-center px-4 py-2 border-b">
+      <button
+        onClick={closeModal}
+        className="text-gray-500 hover:text-gray-700 text-xl font-bold"
+      >
+        X
+      </button>
+      <h3 className="text-lg font-bold text-center flex-1">{title}</h3>
+    </div>
+  );
+}
+
+interface ModalHeaderProps {
+  title: string;
+  closeModal: () => void;
+}
